requirements.txt
bcrypt==4.3.0
cryptography==44.0.2
fastapi==0.115.11
passlib==1.7.4
psycopg2==2.9.10
pydantic-settings==2.8.1
python-jose==3.4.0
sqlmodel==0.0.24
uvicorn==0.34.0
